Use faux bold to prevent wrapping in table of contents
When using `font-weight:bold` for the active item in the table of contents,
the width of the item would change and as a result could sometimes wrap. To
prevent this, we now use `-webkit-text-stroke` to simulate a bold font. It
doesn't look quite as nice, but will do for now.