Add postbuild task to generate a sitemap
Certain people might be interested in using a sitemap with re/blog in
order to improve its discoverability/indexability by search engines. This
shows how it could potentially be done using a simple python script and
a postbuild task. The task will currently only work properly when using
`dotnet publish`, due to the hard-coded paths used in the script.